Lots of work moving from 0.5 to 0.85
Old calculator only really hinted at Nyads and did not have the division fields built in yet. This one does now, but I'm still working on big parts of it. Don't expect the menus to work reliably yet. (if you see an old copyright on a code file, it hasn't been updated yet.)