working off some ideas
and re-arranging things in preparation for support of complex numbers in the calculator. There are a lot of small things to move around because of an over-dependence on RealF as the cladosF magnitude.

I've also renamed a number of methods, but that's mostly internal stuff so I can follow my own code short of writing a book to do it.