Protonumber moved
Moving protonumber from Foot to Algebra means the panels in the calculator have to look a little deeper into nyads to find Cardinals and DivFields for display purposes.

The 'scale' event is obviously impacted too since a new DivField is generated based on what is seen in the FieldBar. The foundation of that DivField is now in an Algebra instead of being up in the nyad.