monad not copying correctly
Stepping back to fix clados. This is a consequence of swithing to
unordered maps for structures within monad. The calculator was assuming
coefficients were in order when working with arrays of them. They no
longer are, so calculator must work with blade/coefficient pairs.

That means some of Monad's constructors have to be updated to avoid
arrays and prefer maps or Scale itself.